services/agents/requirements.txt
fastapi>=0.124.1
uvicorn[standard]>=0.34.0
google-adk==1.14.0
google-cloud-aiplatform
google-auth
httpx
mcp>=1.0.0
